Significantly Improved Performance (When using fences)
Due to reasons not currently understood fully adjusting the number of iterations of the for loop when calling the findQueueFamilies function drastically improves the efficiency of the draw call by approximately 30-40%.

This is presumed to be due to the reduced number of Queues that are enumerated and added to the pipeline, which also reduced the degree of waits required by fences when/during frame synchronisation, and not necessarily due to a graphical enhancement, optimisation or refinement of any kind.

The memory alloctaion model has also been changed to use JEMalloc instead of LibCstd and some of the default MemoryStack handling previously used which as a result of said refactoring may also effect frametime behaviour and consistency.