Further improvements to rendering and added Rudimentary memory Tracking
The prior improvements to rendering were eventually found to be due to how queue families are assigned with Vulkan and handled in hardware; previously the Graphics and PresentQueue used implemented Queue families 0 and 2 respectively, which on Nvidia GPUs implies that the former is capable of Graphics and presenting while the latter can only handle present only, leading to the graphical artefacts if Fences are disabled. However if both queues are assigned to family 0 the 2nd presentQueue becomes capable of handling graphics as well as present, allowing for both queues to be assigned command buffers concurrently, apparently without need for additional synchronisation. Allowing for fences to be removed entirely from the main draw call. Why this is seemingly possible and/or if their are additional issues such as inadvertent frame duplication e.g. are still not currently known or fully understood currently.