Further Optimised Handle use to primitives
Found that Limitations with Inability to properly pass by reference with java could be superseded by using said handle as a long instead of a single array and passing it as  argument to callPPPPI instead of returning written address of callPPPI, allowing the handle to be used depsite bieng Final/Const and being passed by value instead of exploiting pass by reference properties of some object arguments with Function Bodies/Argument Lists e.g.
Also found that to emulate array access and theerfore teh pass by reference capability, memgetlong was needed to allow the reference of the Primitive/Long to be obtained which should have a lower overhead than Null-checked Array Access as is per the Default in Java itself as well as reduced heap allocation/overhead and potentially better memory layout.

The possible drawback to this is that some longs/Handles had to be allocated with JEMalloc's Malloc and/or calloc calls which may negatively effect the memory layout's contiguity and cause potential performance Persistent/retained Degradation/Regressions.